+++ swift_browser_ui_frontend/src/components/BrowserNavbar.vue
@@ -38,7 +38,9 @@
                 }"
                 class="navbar-item"
                 @click.native.stop="changeActive(item)"
-              >{{ item.name }}</router-link>
+              >
+                {{ item.name }}
+              </router-link>
             </div>
           </div>
           <div