+++ tests/cypress/integration/userInfo.cy.js
@@ -11,7 +11,8 @@ describe("Login and log out a user", function () {
     })
 
     it("should login user with Finnish to username + password and remember the selection", () => {
-        cy.get('select').select('Suomeksi')
+        cy.visit(Cypress.config().baseUrl)
+        cy.contains("In English").click({ position: "top" }).shadow().find("li").contains("Suomeksi").click()
         cy.login(' Kirjaudu SSO:ta käyttäen ')
         cy.get('.buttons > .router-link-exact-active').should(($browse) => {
             expect($browse).to.have.length(1)